refactor: mock server 패키지 분리
