+++ app/src/main/java/woowacourse/shopping/mockserver/ProductStorage.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.mockserver
 
 import woowacourse.shopping.domain.model.Price
 import woowacourse.shopping.domain.model.Product
